requirements.txt
fastapi
uvicorn
streamlit
yfinance
pandas
numpy
requests
beautifulsoup4
transformers
torch
openai
langchain
python-dotenv
tf-keras
plotly
SpeechRecognition
pyttsx3
